Train and predict gene's predictive model with mediators using DePMA

The function trains a predictive model of a given gene using top mediators as fixed effects and assesses in-sample performance with cross-validation.

trainDeP(
  geneInt,
  snps,
  snpLocs,
  mediator,
  medLocs,
  covariates,
  cisDist = 5e+05,
  qtlTra,
  qtMed,
  h2Pcutoff,
  dimNumeric,
  verbose,
  seed,
  sobel = F,
  nperms = 1000,
  k,
  parallel,
  parType = "no",
  prune,
  windowSize = 50,
  numSNPShift = 5,
  ldThresh = 0.5,
  cores,
  qtlTra_parts,
  qtMed_parts,
  modelDir,
  snpAnnot = NULL
)

Arguments

geneInt
character, identifier for gene of interest

snps
data frame, SNP dosages

snpLocs
data frame, MatrixEQTL locations for SNPs

mediator
data frame, mediator intensities

medLocs
data frame, MatrixEQTL locations for mediators

covariates
data frame, covariates

h2Pcutoff
numeric, P-value cutoff for heritability

seed
integer, random seed for splitting

k
integer, number of training-test splits

parallel
logical, TRUE/FALSE to run glmnet in parallel

prune
logical, TRUE/FALSE to LD prune the genotypes

windowSize
integer, window size for PLINK pruning

numSNPShift
integer, shifting window for PLINK pruning

ldThresh
numeric, LD threshold for PLINK pruning

cores
integer, number of parallel cores

qtlFull
data frame, all QTLs (cis and trans) between mediators and genes

numMed
integer, number of top mediators to include

outputAll
logical, include mediator information

Value

final model for gene along with CV R2 and predicted values